refactor(cuaktask): rename NodeDependenciesType module to NodeDependenciesType
