+++ packages/cuaktask/src/graph/models/domain/LazyNodeDependencies.ts
@@ -1,6 +1,6 @@
 import { BaseNodeDependencies } from './BaseNodeDependencies';
 import { NodeDependencies } from './NodeDependencies';
-import { NodeDependenciesType } from './NodeExecutionOperator';
+import { NodeDependenciesType } from './NodeDependenciesType';
 
 export interface LazyNodeDependencies<TElem>
   extends BaseNodeDependencies<NodeDependenciesType.lazy, TElem> {
