+++ packages/cuaktask/src/graph/models/domain/NodeDependencies.ts
@@ -1,6 +1,6 @@
 import { BaseNodeDependencies } from './BaseNodeDependencies';
 import { LazyNodeDependencies } from './LazyNodeDependencies';
-import { NodeDependenciesType } from './NodeExecutionOperator';
+import { NodeDependenciesType } from './NodeDependenciesType';
 
 export type NodeDependencies<TElem> =
   | BaseNodeDependencies<NodeDependenciesType.and, TElem>
